refactor(open-mpm): split memory/redb_usearch + graph into submodules (#366)
